server/requirements.txt
fastapi
uvicorn
pydantic
Pillow
numpy
python-multipart
git+https://github.com/Deepshift/DeepCreamPy.git@master#egg=lib-cream-py&subdirectory=lib-cream-py